API/requirements.txt
fastapi
sqlalchemy
asyncpg
python-multipart
minio
python-dotenv